calculate index 50

+++ src/main/java/cc/xpbootcamp/warmup/fibonacci/FibonacciCalculator.java
@@ -1,7 +1,7 @@
 package cc.xpbootcamp.warmup.fibonacci;
 
 public class FibonacciCalculator {
-	public static int calculate(int index) {
+	public static long calculate(int index) {
 		if (index == 1 || index == 2) {
 			return 1;
 		}
